requirements.txt
streamlit
google-generativeai
Jinja2
streamlit-ace
firebase-admin
Werkzeug
python-docx
toml